Use explicit toUInt() for andR/orR; convert UInt(width=1) to Bool(); allow HW testing.

+++ src/main/scala/regfile.scala
@@ -33,7 +33,7 @@ class RouterRegFile(parms: Parameters) extends Module(parms) {
 
 
 	io.full := andR(regFileValid.toBits())
-	io.readValid := (writePointer =/= readPointer) && orR(regFileValid.toBits())
+	io.readValid := (writePointer =/= readPointer) && orR(regFileValid.toBits().toUInt())
 
 	when (io.writeEnable && !regFileValid(writePointer)) {
 		regFile(writePointer) := io.writeData
