+++ src/main/scala/router.scala
@@ -812,7 +812,7 @@ class SimpleVCRouter(parms: Parameters) extends VCRouter(parms) {
         // ------------------- 
 
         //State Machine
-        routerOutputStateMgmt(i).io.swAllocGranted  := orR(Vec (readyToXmit.map( a=> a(i))).toBits)  //This should be routed from the input side
+        routerOutputStateMgmt(i).io.swAllocGranted  := orR(Vec (readyToXmit.map( a=> a(i))).toBits().toUInt())  //This should be routed from the input side
         routerOutputStateMgmt(i).io.creditsAvail    := outCredits(deqVCPort) //&& ~almostOuts(deqVCPort)
         // -------------------
